feat: create the nav bar.
it contains 3 stars each star is a span with three classes 'star','fa',
and 'fa-star'.
and there is a section for moves while playing it's value will change.
the timer will start count after the first ckick it is formated as :
00:00 --> minutes:seconds .
the replay just for replay the game whenever you want.