feat: create tow functions to choose cards randomly
pickCards(array) : is a function to choose 8 cards randomly from an
array that will be passed to the function as a parameter
and after choosing the cards they will be doubled
because we have 16 cards.
the 16 chosen cards will be returned as an array called : pickedArr.

shuffle(array) : is a function to shuffle a given array and return it.
so after we create the pickedArr we will pass it to the shuffle()
so the cards will be randomly shuffled.