feat : create the startPlaying(), stopPlaying(), checkForWin(), and replay click events
the startPlaying() function is to start the game and make the clicks
counted and check if cards are the same or not and start animations.

stopPlaying() on the other hand for stop playing and clear every
variable and reset it to it's initial state.

checkForWin() function to check if the player won the game or not.

the replay click event will cause the game to stop and start again.